+++ example/lib/pokemon/widgets/pokemon_item_widget.dart
@@ -1,25 +1,30 @@
 import 'package:examplecube/pokemon/repository/model/pokemon.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 class PokemonItemWidget extends StatefulWidget {
-  final Pokemon item;
+  final Pokemon? item;
 
-  final Function(Pokemon) onClick;
+  final Function(Pokemon?)? onClick;
 
-  const PokemonItemWidget({Key key, this.item, this.onClick}) : super(key: key);
+  const PokemonItemWidget({Key? key, this.item, this.onClick})
+      : super(key: key);
   @override
   _PokemonItemWidgetState createState() => _PokemonItemWidgetState();
 }
 
-class _PokemonItemWidgetState extends State<PokemonItemWidget> with TickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<Offset> _animationSlide;
+class _PokemonItemWidgetState extends State<PokemonItemWidget>
+    with TickerProviderStateMixin {
+  late AnimationController _controller;
+  late Animation<Offset> _animationSlide;
 
   @override
   void initState() {
-    _controller = AnimationController(vsync: this, duration: Duration(seconds: 1));
+    _controller =
+        AnimationController(vsync: this, duration: Duration(seconds: 1));
     _animationSlide = Tween(begin: Offset(2.0, 0.0), end: Offset(0.0, 0.0))
-        .animate(CurvedAnimation(parent: _controller, curve: Curves.decelerate));
+        .animate(
+            CurvedAnimation(parent: _controller, curve: Curves.decelerate));
     _controller.forward(from: 0.0);
     super.initState();
   }
@@ -40,15 +45,15 @@ class _PokemonItemWidgetState extends State<PokemonItemWidget> with TickerProvid
   ListTile _getListTile() {
     return new ListTile(
       contentPadding: const EdgeInsets.all(10),
-      leading: _getLeadingWidget(widget.item.thumbnailImage),
-      title: _getTittleWidget(widget.item.name),
+      leading: _getLeadingWidget(widget.item!.thumbnailImage!),
+      title: _getTittleWidget(widget.item!.name!),
       subtitle: Text(
-        widget.item.description,
+        widget.item!.description!,
         maxLines: 3,
       ),
       onTap: () {
         if (widget.onClick != null) {
-          widget.onClick(widget.item);
+          widget.onClick!(widget.item);
         }
       },
     );
@@ -58,7 +63,7 @@ class _PokemonItemWidgetState extends State<PokemonItemWidget> with TickerProvid
     return Container(
       width: 50,
       height: 50,
-      child: Image.network(url),
+      child: Image.network(_getUrl(url)),
     );
   }
 
@@ -76,5 +81,13 @@ class _PokemonItemWidgetState extends State<PokemonItemWidget> with TickerProvid
   void dispose() {
     _controller.dispose();
     super.dispose();
+  }
+
+  String _getUrl(String url) {
+    String CORS = '';
+    if (kIsWeb) {
+      CORS = 'https://cors-anywhere.herokuapp.com/';
+    }
+    return '$CORS$url';
   }
 }