+++ lib/src/util/extensions/build_context_extensions.dart
@@ -1,13 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
+import '../../../cubes.dart';
+import '../../cube.dart';
+
 extension BuildContextExtensions on BuildContext {
-  Future<T> goTo<T extends Object>(
+  Future<T?> goTo<T extends Object?>(
     Widget widget, {
-    RouteSettings settings,
+    RouteSettings? settings,
     bool fullscreenDialog = false,
   }) {
-    return Navigator.push<T>(
+    return Navigator.push<T?>(
       this,
       MaterialPageRoute(
         builder: (context) => widget,
@@ -17,9 +20,9 @@ extension BuildContextExtensions on BuildContext {
     );
   }
 
-  Future<T> goToReplacement<T extends Object, TO extends Object>(
+  Future<T?> goToReplacement<T extends Object?, TO extends Object?>(
     Widget widget, {
-    RouteSettings settings,
+    RouteSettings? settings,
     bool fullscreenDialog = false,
   }) {
     return Navigator.pushReplacement<T, TO>(
@@ -32,10 +35,10 @@ extension BuildContextExtensions on BuildContext {
     );
   }
 
-  Future<T> goToAndRemoveUntil<T extends Object>(
+  Future<T?> goToAndRemoveUntil<T extends Object?>(
     Widget widget,
     RoutePredicate predicate, {
-    RouteSettings settings,
+    RouteSettings? settings,
     bool fullscreenDialog = false,
   }) {
     return Navigator.pushAndRemoveUntil<T>(
@@ -49,7 +52,7 @@ extension BuildContextExtensions on BuildContext {
     );
   }
 
-  void pop<T>([T result]) {
+  void pop<T>([T? result]) {
     Navigator.pop(this, result);
   }
 
@@ -63,4 +66,12 @@ extension BuildContextExtensions on BuildContext {
 
   ThemeData get theme => Theme.of(this);
   ScaffoldState get scaffold => Scaffold.of(this);
+
+  void showSnackBar(SnackBar snackBar) {
+    ScaffoldMessenger.of(this).showSnackBar(snackBar);
+  }
+
+  C? getCube<C extends Cube>() => Cubes.of<C>(this);
+
+  Object? get arguments => ModalRoute.of(this)?.settings.arguments;
 }