+++ lib/src/util/state_mixin.dart
@@ -17,28 +17,28 @@ mixin StateMixin<T extends StatefulWidget> on State<T> {
 /// Mixin to user Cube in StatefulWidget
 mixin CubeStateMixin<T extends StatefulWidget, C extends Cube> on State<T> {
   /// Cube that will be used
-  C cube;
+  C? cube;
 
   /// Initial data used in `cube.onReady`
-  Object get initData => null;
+  Object? get initData => null;
 
   @override
   void initState() {
     cube = cube ?? Cubes.getDependency();
-    cube.addOnActionListener(_innerOnAction);
-    WidgetsBinding.instance.addPostFrameCallback(_ready);
+    cube?.addOnActionListener(_innerOnAction);
+    WidgetsBinding.instance?.addPostFrameCallback(_ready);
     super.initState();
   }
 
   @override
   void dispose() {
-    cube.dispose();
+    cube?.dispose();
     removeOnActionListener();
     super.dispose();
   }
 
   /// Remove action listeners.
-  void removeOnActionListener() => cube.removeOnActionListener(_innerOnAction);
+  void removeOnActionListener() => cube?.removeOnActionListener(_innerOnAction);
 
   void _innerOnAction(C cube, CubeAction action) => onAction(action);
 
@@ -46,7 +46,7 @@ mixin CubeStateMixin<T extends StatefulWidget, C extends Cube> on State<T> {
   void onAction(CubeAction action);
 
   void _ready(_) {
-    var data = initData ?? ModalRoute.of(context)?.settings?.arguments;
-    cube.onReady(data);
+    var data = initData ?? ModalRoute.of(context)?.settings.arguments;
+    cube?.onReady(data);
   }
 }